HTML output + basic `type` + `dispatch` annotations
The type annotations are still fairly buggy. Sometimes they seem to
magically "transport" types backwards-in-time, or the annotation will
skip something that it seems should "obviously" have a type associated
with it.

Dispatch annotations are a proof-of-concept only. We really want to
instead collect this dispatch information straight from inference, if
we can.

Also does not support kwargs/varargs yet